+++ test/replay/btcstaking_test.go
@@ -605,7 +605,7 @@ func TestBadUnbondingFeeParams(t *testing.T) {
 		Params:    p,
 	}
 	msgToSend := d.NewGovProp(&prop)
-	d.SendTxWithMessagesSuccess(t, d.SenderInfo, defaultGasLimit, defaultFeeCoin, msgToSend)
+	d.SendTxWithMessagesSuccess(t, d.SenderInfo, DefaultGasLimit, defaultFeeCoin, msgToSend)
 
 	txResults := d.GenerateNewBlockAssertExecutionFailure()
 	require.Len(t, txResults, 1)
@@ -656,7 +656,6 @@ func TestPostRegistrationDelegation(t *testing.T) {
 	activeDelegations := driver.GetActiveBTCDelegations(t)
 	require.Len(t, activeDelegations, 1)
 }
-
 
 func containsEvent(events []abci.Event, eventType string) bool {
 	for _, event := range events {
@@ -754,4 +753,113 @@ func TestAcceptSlashingTxAsUnbondingTx(t *testing.T) {
 
 	unbondedDelegations := driver.GetUnbondedBTCDelegations(t)
 	require.Len(t, unbondedDelegations, 1)
+}
+
+func TestSlashingFpWithManyMulistakedDelegations(t *testing.T) {
+	tmpGas := DefaultGasLimit
+	DefaultGasLimit = uint64(10_000_000)
+	defer func() {
+		DefaultGasLimit = tmpGas
+	}()
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	d := NewBabylonAppDriverTmpDir(r, t)
+	d.GenerateNewBlockAssertExecutionSuccess()
+
+	covSender := d.CreateCovenantSender()
+	bbnFp := d.CreateNFinalityProviderAccounts(1)[0]
+	numStakers := 50
+	stakers := d.CreateNStakerAccounts(numStakers)
+	d.GenerateNewBlockAssertExecutionSuccess()
+
+	bbnFp.RegisterFinalityProvider("")
+	d.GenerateNewBlockAssertExecutionSuccess()
+
+	bbnFp.CommitRandomness()
+
+	currentEpochNumber := d.GetEpoch().EpochNumber
+	d.ProgressTillFirstBlockTheNextEpoch()
+	d.FinializeCkptForEpoch(currentEpochNumber)
+
+	d.GenerateNewBlockAssertExecutionSuccess()
+
+	stk2 := stakers[0]
+	// send one btc delegation just to have voting power in the fp
+	fps := []*bbn.BIP340PubKey{bbnFp.BTCPublicKey()}
+	d.SendAndVerifyNDelegations(t, stk2, covSender, fps, 1)
+
+	d.ActivateVerifiedDelegations(1)
+	d.GenerateNewBlockAssertExecutionSuccess()
+
+	// fp is activated
+	activationHeight := d.GetActivationHeight(d.t)
+	require.Greater(d.t, activationHeight, uint64(0))
+
+	// create consumers
+	const consumerID1 = "consumer1"
+	const consumerID2 = "consumer2"
+	ctx := d.Ctx().WithIsCheckTx(false)
+	OpenChannelForConsumer(ctx, d.App, consumerID1)
+	OpenChannelForConsumer(ctx, d.App, consumerID2)
+
+	// 2. Register consumers
+	consumer1 := d.RegisterConsumer(r, consumerID1)
+	consumer2 := d.RegisterConsumer(r, consumerID2)
+	require.NotNil(t, consumer1, consumer2)
+
+	// 3. Create finality providers for each consumer
+	fpsCons1 := d.CreateFinalityProviderForConsumer(consumer1)
+	fpsCons2 := d.CreateFinalityProviderForConsumer(consumer2)
+	require.NotNil(t, fpsCons1, fpsCons2)
+
+	d.GenerateNewBlockAssertExecutionSuccess()
+	fps = []*bbn.BIP340PubKey{bbnFp.BTCPublicKey(), fpsCons1.BTCPublicKey(), fpsCons2.BTCPublicKey()}
+
+	d.MintNativeTo(t, covSender.Address(), 10000000_000000)
+
+	// creates 200 btc delegations to slash it
+	batchSize := 4
+	totalActiveDels := len(d.GetActiveBTCDelegations(d.t))
+	for _, stk := range stakers {
+		d.MintNativeTo(t, stk.Address(), 10000000_000000)
+		d.SendAndVerifyNDelegations(t, stk, covSender, fps, batchSize)
+
+		d.GenerateNewBlockAssertExecutionSuccess()
+		d.GenerateNewBlockAssertExecutionSuccess()
+
+		verifiedDelegations := d.GetVerifiedBTCDelegations(t)
+		require.Equal(t, len(verifiedDelegations), batchSize)
+
+		d.ActivateVerifiedDelegations(batchSize)
+
+		d.GenerateNewBlockAssertExecutionSuccess()
+		activeDels := d.GetActiveBTCDelegations(d.t)
+
+		totalActiveDels += batchSize
+		require.Equal(t, len(activeDels), totalActiveDels)
+	}
+	// 200 dels + bbn del
+	require.Equal(t, totalActiveDels, (batchSize*numStakers)+1)
+
+	// unjails the fp so he can vote
+	fpBtcPk := bbnFp.BTCPublicKey()
+	fpBbn, err := d.App.BTCStakingKeeper.GetFinalityProvider(d.Ctx(), *fpBtcPk)
+	require.NoError(t, err)
+
+	fpBbn.Jailed = false
+	d.App.BTCStakingKeeper.SetFinalityProvider(d.Ctx(), fpBbn)
+
+	d.GenerateNewBlockAssertExecutionSuccess()
+	bbnFp.CastVote(activationHeight)
+	d.GenerateNewBlockAssertExecutionSuccess()
+
+	// slash it
+	bogusHash := datagen.GenRandomByteArray(r, 32)
+	txRes := bbnFp.CastVoteForHash(activationHeight, bogusHash)
+	require.LessOrEqual(t, txRes.GasUsed, int64(100_000))
+	res := d.GenerateNewBlockAssertExecutionSuccessWithResults()
+	require.NotEmpty(t, res)
+
+	fpBbn, err = d.App.BTCStakingKeeper.GetFinalityProvider(d.Ctx(), *fpBtcPk)
+	require.NoError(t, err)
+	require.True(t, fpBbn.IsSlashed())
 }