+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+- [#1525](https://github.com/babylonlabs-io/babylon/pull/1525) Remove `PropagateFPSlashingToConsumers` call from slashing logic at send
+finality signature and `SlashedBTCDelegation` structure.
 - [#1229](https://github.com/babylonlabs-io/babylon/pull/1229) Wire up EVM.
 
 ## v3.0.0-rc.0